[RWA-0043] increased service layer functionality.

+++ src/main/java/com/october/to/finish/restaurantwebapp/service/impl/UserServiceImpl.java
@@ -1,4 +1,5 @@
 package com.october.to.finish.restaurantwebapp.service.impl;
+
 import com.october.to.finish.restaurantwebapp.dao.CreditCardDAO;
 import com.october.to.finish.restaurantwebapp.dao.UserDAO;
 import com.october.to.finish.restaurantwebapp.exceptions.DAOException;
@@ -17,11 +18,11 @@ public class UserServiceImpl implements UserService {
     private static final Logger LOGGER = LogManager.getLogger(UserServiceImpl.class);
     private static final String NULL_USER_DAO_EXC = "[UserService] Can't create UserService with null input UserDAO";
     private static final String NULL_CARD_DAO_EXC = "[UserService] Can't create UserService with null input CreditCardDAO";
-    private static final String NULL_USER_INPUT_EXC = "[UserService] Can't operate null input User!";
+    private static final String NULL_USER_INPUT_EXC = "[UserService] Can't operate null input!";
     private static final String REGISTERED_EMAIL_EXC =
             "[UserService] User with given email is already registered! (email: {})";
-    UserDAO userDAO;
-    CreditCardDAO creditCardDAO;
+    private final UserDAO userDAO;
+    private CreditCardDAO creditCardDAO;
 
     public UserServiceImpl(UserDAO userDAO) {
         if (userDAO == null) {
@@ -46,6 +47,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public boolean save(User user) throws ServiceException {
+        if (user == null) {
+            LOGGER.error(NULL_USER_INPUT_EXC);
+            throw new IllegalArgumentException(NULL_USER_INPUT_EXC);
+        }
         try {
             return checkAndSave(user);
         } catch (SQLException e) {
@@ -56,10 +61,6 @@ public class UserServiceImpl implements UserService {
     }
 
     private boolean checkAndSave(User user) throws ServiceException, SQLException {
-        if (user == null) {
-            LOGGER.error(NULL_USER_INPUT_EXC);
-            throw new IllegalArgumentException(NULL_USER_INPUT_EXC);
-        }
         userDAO.getConnection().setAutoCommit(false);
         try {
             if (userDAO.findByEmail(user.getEmail()).getId() != 0) {
@@ -85,6 +86,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User findById(long userId) throws ServiceException {
+        if (userId < 1) {
+            LOGGER.error(NULL_USER_INPUT_EXC);
+            throw new IllegalArgumentException(NULL_USER_INPUT_EXC);
+        }
         try {
             return userDAO.findById(userId);
         } catch (DAOException e) {
@@ -96,6 +101,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User findByEmail(String eMail) throws ServiceException {
+        if (eMail == null) {
+            LOGGER.error(NULL_USER_INPUT_EXC);
+            throw new IllegalArgumentException(NULL_USER_INPUT_EXC);
+        }
         try {
             return userDAO.findByEmail(eMail);
         } catch (DAOException e) {
@@ -116,18 +125,26 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public boolean update(User user) throws ServiceException {
+    public boolean update(long userId, User user) throws ServiceException {
+        if (userId < 1 || user == null) {
+            LOGGER.error(NULL_USER_INPUT_EXC);
+            throw new IllegalArgumentException(NULL_USER_INPUT_EXC);
+        }
         try {
-            return userDAO.update(user.getId(), user);
+            return userDAO.update(userId, user);
         } catch (DAOException e) {
             LOGGER.error("[UserService] An exception occurs while updating User. (id: {}). Exc: {}"
-                    , user.getId(), e.getMessage());
+                    , userId, e.getMessage());
             throw new ServiceException(e.getMessage(), e);
         }
     }
 
     @Override
     public void delete(long userId) throws ServiceException {
+        if (userId < 1) {
+            LOGGER.error(NULL_USER_INPUT_EXC);
+            throw new IllegalArgumentException(NULL_USER_INPUT_EXC);
+        }
         try {
             userDAO.delete(userId);
         } catch (DAOException e) {
@@ -139,6 +156,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void saveCreditCard(User user, CreditCard creditCard) throws ServiceException {
+        if (user == null || creditCard == null) {
+            LOGGER.error(NULL_USER_INPUT_EXC);
+            throw new IllegalArgumentException(NULL_USER_INPUT_EXC);
+        }
         try {
             creditCardDAO.save(user.getId(), creditCard);
         } catch (DAOException e) {
@@ -150,6 +171,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public CreditCard findCreditCard(User user) throws ServiceException {
+        if (user == null) {
+            LOGGER.error(NULL_USER_INPUT_EXC);
+            throw new IllegalArgumentException(NULL_USER_INPUT_EXC);
+        }
         try {
             return creditCardDAO.findByUser(user.getId());
         } catch (DAOException e) {
@@ -161,7 +186,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void updateCreditCard(User user, CreditCard creditCard) throws ServiceException {
-        if (user.getCreditCard() == null) {
+        if (user.getCreditCard().getCardNumber() == null) {
             throw new IllegalArgumentException("[UserService] Can't update null CreditCard");
         }
         try {
@@ -175,6 +200,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteCreditCard(User user) throws ServiceException {
+        if (user == null) {
+            LOGGER.error(NULL_USER_INPUT_EXC);
+            throw new IllegalArgumentException(NULL_USER_INPUT_EXC);
+        }
         try {
             creditCardDAO.deleteByUserId(user.getId());
         } catch (DAOException e) {
